keyring: remove the file param
Being able to provide the dir, filename, and full file path makes the
logic too complicated for users of this defined type. We should only
allow providing the filename and dir, and then compute the full path
from those. Otherwise, someone could pass all three parameters, in which
case the filename and dir would be ignored, which could be confusing.